update overall doc, operator, and apis
