+++ .readthedocs.yml
@@ -18,3 +18,5 @@ sphinx:
 python:
   install:
     - requirements: requirements-doc.txt
+
+
