+++ braintaichi/_primitive/_mlir_translation_rule.py
@@ -27,7 +27,6 @@ import shutil
 from functools import partial
 from typing import Any, Sequence, Union
 
-import jax.core
 import numpy as np
 import taichi as ti
 from jax.interpreters import mlir
@@ -35,16 +34,24 @@ from jax.lib import xla_client
 from jaxlib.hlo_helpers import custom_call
 
 from ._batch_utils import _shape_to_layout
-from . import cpu_ops
-for _name, _value in cpu_ops.registrations().items():
-  xla_client.register_custom_call_target(_name, _value, platform="cpu")
+
+# --- REGISTER CUSTOM CALL TARGETS on CPU platforms ###
 try:
-  from . import gpu_ops
+  from braintaichi import cpu_ops # noqa
+
+  for _name, _value in cpu_ops.registrations().items():
+    xla_client.register_custom_call_target(_name, _value, platform="cpu")
+except ImportError:
+  cpu_ops = None
+
+# --- REGISTER CUSTOM CALL TARGETS on GPU platforms ###
+try:
+  from braintaichi import gpu_ops # noqa
+
   for _name, _value in gpu_ops.registrations().items():
     xla_client.register_custom_call_target(_name, _value, platform="gpu")
 except ImportError:
   gpu_ops = None
-
 
 taichi_cache_path = None
 
@@ -156,7 +163,7 @@ def is_metal_supported():
 
 # --- VARIABLES ###
 home_path = get_home_dir()
-kernels_aot_path = os.path.join(home_path, '.brainpy', 'kernels')
+kernels_aot_path = os.path.join(home_path, '.braintaichi', 'kernels')
 is_metal_device = is_metal_supported()
 
 
@@ -201,9 +208,11 @@ def _array_to_field(dtype, shape) -> Any:
   elif dtype == np.float64:
     dtype = ti.float64
   else:
-    raise NotImplementedError(f'Currently we do not support dtype {dtype} in Taichi. '
-                              f'If you think it is necessary, please open an issue at '
-                              f'https://github.com/brainpy/BrainPy/issues/new')
+    raise NotImplementedError(
+      f'Currently we do not support dtype {dtype} in Taichi. '
+      f'If you think it is necessary, please open an issue at '
+      f'https://github.com/chaoming0625/braintaichi/issues/new'
+    )
   return ti.field(dtype=dtype, shape=shape)
 
 
@@ -382,15 +391,6 @@ def _preprocess_kernel_call_gpu(
   return opaque
 
 
-def _XlaOp_to_ShapedArray(c, xla_op):
-  xla_op = c.get_shape(xla_op)
-  return jax.core.ShapedArray(xla_op.dimensions(), xla_op.element_type())
-
-
-def _mlir_to_ShapedArray(c, op):
-  return op
-
-
 def _kernel_to_code(kernel, abs_ins, abs_outs, platform):
   codes = f'[taichi {platform} kernel]\n' + get_source_with_dependencies(kernel)
   codes += '\n[ins]: {}'.format("-".join([f'{v.dtype}[{v.shape}]' for v in abs_ins]))
@@ -479,10 +479,22 @@ def _taichi_mlir_gpu_translation_rule(kernel, c, *ins, **kwargs):
 
 
 def register_taichi_aot_mlir_cpu_translation_rule(primitive, cpu_kernel):
+  if cpu_ops is None:
+    raise RuntimeError(
+      'The CPU kernels do not build correctly. '
+      'Please check the installation of braintaichi.'
+    )
+
   rule = partial(_taichi_mlir_cpu_translation_rule, cpu_kernel)
   mlir.register_lowering(primitive, rule, platform='cpu')
 
 
 def register_taichi_aot_mlir_gpu_translation_rule(primitive, gpu_kernel):
+  if gpu_ops is None:
+    raise RuntimeError(
+      'The GPU kernels are not supported on this device. '
+      'Please install the GPU supported version of braintaichi.'
+    )
+
   rule = partial(_taichi_mlir_gpu_translation_rule, gpu_kernel)
   mlir.register_lowering(primitive, rule, platform='gpu')
