+++ braintaichi/_sparseop/_sparse_csrmm.py
@@ -25,41 +25,8 @@ from jax import numpy as jnp
 from jax.experimental.sparse import csr
 from jax.interpreters import ad
 
-from ._base import XLACustomOp
-from ._batch_utils import register_general_batching
-
-__all__ = [
-  'csrmm',
-]
-
-
-def csrmm(
-    data: jax.typing.ArrayLike | u.Quantity,
-    indices: jax.typing.ArrayLike,
-    indptr: jax.typing.ArrayLike,
-    matrix: jax.typing.ArrayLike,
-    *,
-    shape: Tuple[int, int],
-    transpose: bool = False,
-):
-  """
-  Product of CSR sparse matrix and a dense matrix.
-
-  Args:
-      data : array of shape ``(nse,)``.
-      indices : array of shape ``(nse,)``
-      indptr : array of shape ``(shape[0] + 1,)`` and dtype ``indices.dtype``
-      B : array of shape ``(shape[0] if transpose else shape[1], cols)`` and
-      dtype ``data.dtype``
-      shape : length-2 tuple representing the matrix shape
-      transpose : boolean specifying whether to transpose the sparse matrix
-      before computing.
-
-  Returns:
-      C : array of shape ``(shape[1] if transpose else shape[0], cols)``
-      representing the matrix-matrix product.
-  """
-  return raw_csrmm_taichi(data, indices, indptr, matrix, shape=shape, transpose=transpose)[0]
+from braintaichi._primitive._batch_utils import register_general_batching
+from braintaichi._primitive._xla_custom_op import XLACustomOp
 
 
 def raw_csrmm_taichi(
