+++ docs/apis/event-operators.rst
@@ -1,0 +1,16 @@
+Event-driven Computation Operators
+==================================
+
+.. currentmodule:: braintaichi
+.. automodule:: braintaichi
+
+
+.. autosummary::
+   :toctree: generated/
+   :nosignatures:
+   :template: classtemplate.rst
+
+    event_csrmv
+    event_csrmm
+
+
