+++ docs/apis/jitconn-operators.rst
@@ -1,0 +1,20 @@
+JIT Connectivity Operators
+==========================
+
+.. currentmodule:: braintaichi
+.. automodule:: braintaichi
+
+
+.. autosummary::
+   :toctree: generated/
+   :nosignatures:
+   :template: classtemplate.rst
+
+   mv_prob_homo
+   mv_prob_uniform
+   mv_prob_normal
+   event_mv_prob_homo
+   event_mv_prob_uniform
+   event_mv_prob_normal
+
+
