+++ docs/apis/operator-registration.rst
@@ -1,0 +1,17 @@
+Operator Registration Routine
+=============================
+
+.. currentmodule:: braintaichi
+.. automodule:: braintaichi
+
+
+.. autosummary::
+   :toctree: generated/
+   :nosignatures:
+   :template: classtemplate.rst
+
+    XLACustomOp
+    defjvp
+    register_general_batching
+
+
