+++ docs/apis/sparse-operators.rst
@@ -1,0 +1,17 @@
+Sparse Computation Operators
+============================
+
+.. currentmodule:: braintaichi
+.. automodule:: braintaichi
+
+
+.. autosummary::
+   :toctree: generated/
+   :nosignatures:
+   :template: classtemplate.rst
+
+    coomv
+    csrmv
+    csrmm
+
+
